requirements.txt
pyyaml
fastapi==0.95.1
uvicorn
py3langid
setuptools-scm
openai==0.27.8
opencv-python==4.8.0.76
gradio
httpx[socks]==0.25.0
astunparse
langchain==0.0.157
scikit-learn
pydantic==1.10.7
typing-extensions==4.5.0
typing-inspect==0.8.0
colorlog
rapidfuzz
spacy
seaborn
colorama==0.4.6
tiktoken==0.5.1
Mesa==2.2.4
Mesa-Viz-Tornado==0.1.3
rank_bm25==0.2.2
textblob==0.18.0.post0